line-bot/requirements.txt
fastapi
uvicorn
line-bot-sdk
anthropic
python-dotenv
